+++ tests/Rules/Data/UselessConstructs/UselessWithFunctionCall.php
@@ -1,0 +1,19 @@
+<?php
+declare(strict_types=1);
+
+namespace Rules\Data\UselessConstructs;
+
+use function with;
+
+class UselessWithFunctionCall
+{
+    public function foo(): string
+    {
+        return with('foo');
+    }
+
+    public function bar(): string
+    {
+        return with('bar', null);
+    }
+}