+++ CHANGELOG.md
@@ -54,6 +54,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Improvements
 * feat: add cursorPaginate() support by @BramVanBerkel in https://github.com/nunomaduro/larastan/pull/1198
 * feat: add App::make() support by @PrinsFrank in https://github.com/nunomaduro/larastan/pull/1222
+* feat: add Rule checking for useless 'with()' function calls by @PrinsFrank
 
 ### Fixes
 * fix: inverted check in the ModelRuleHelper by @mad-briller in https://github.com/nunomaduro/larastan/pull/1207